Adds optional preprocessing step for GWT specific version of the library
Adds preprocessor directives which comment parts of the code
conditionaly (for now only Coordinate#clone())

For now preprocessor doesn't support test source folders, so tests will
fail because of missing method. (Only when preprocessing is ON)

Replaces Comparable<Coordinate with Object> to limit API changes and
remove warnings. 

Modifies the comment since it was treated as invalid preprocessor
directive (//*c...)