Modifies how gwt package is built
Now jts-core is built only with preprocessing enabled. Then when jts-gwt
is built it compiles its own classes and then unpacks gwt specific core
with some exclusions. So now it's possible to provide gwt specific
implementation of some classes (which have to be excluded from copying).

First class which overrides jvm impl is NumberFormat which was moved
from jtsemul.